+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v1.6.0 
+*2014-04-06*
+
 **MayhemYDG**:
 - Thread and post hiding changes:
  - The posts' menu now has a label entry listing the reasons why a post got hidden or highlighted.